+++ app/src/main/AndroidManifest.xml
@@ -16,16 +16,16 @@
         android:theme="@style/Theme.Movie"
         tools:targetApi="31">
         <activity
-            android:name=".view.booking.seat.BookingSeatActivity"
+            android:name=".ui.seat.BookingSeatActivity"
             android:exported="false" />
         <activity
-            android:name=".view.booking.booking.BookingActivity"
+            android:name=".ui.booking.view.BookingActivity"
             android:exported="false" />
         <activity
-            android:name=".view.booking.complete.BookingCompleteActivity"
+            android:name=".ui.complete.BookingCompleteActivity"
             android:exported="false" />
         <activity
-            android:name=".view.movie.MovieListActivity"
+            android:name=".ui.movielist.view.MovieListActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
