+++ app/src/main/java/woowacourse/movie/ui/complete/BookingCompleteActivity.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.view.booking.complete
+package woowacourse.movie.ui.complete
 
 import android.content.Context
 import android.content.Intent
