+++ app/src/main/java/woowacourse/movie/ui/complete/BookingCompleteContract.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.view.booking.complete
+package woowacourse.movie.ui.complete
 
 import woowacourse.movie.domain.model.BookedTicket
