+++ app/src/main/java/woowacourse/movie/ui/movielist/contract/MovieListContract.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.view.movie
+package woowacourse.movie.ui.movielist.contract
 
 import woowacourse.movie.domain.model.Movie
 import woowacourse.movie.domain.model.MovieListItem
