+++ app/src/main/java/woowacourse/movie/ui/movielist/presenter/MovieListPresenter.kt
@@ -1,17 +1,26 @@
-package woowacourse.movie.view.movie
+package woowacourse.movie.ui.movielist.presenter
 
 import woowacourse.movie.domain.model.Advertisement
 import woowacourse.movie.domain.model.Movie
 import woowacourse.movie.domain.model.MovieListItem
-import woowacourse.movie.domain.model.MovieListItem.AdItem
-import woowacourse.movie.domain.model.MovieListItem.MovieItem
+import woowacourse.movie.ui.movielist.contract.MovieListContract
 
 class MovieListPresenter(
     val movieListView: MovieListContract.View,
 ) : MovieListContract.Presenter {
-    override fun getMovieList(): List<MovieItem> = Movie.DUMMY_MOVIES.map { MovieItem(it) }
+    override fun getMovieList(): List<MovieListItem.MovieItem> =
+        Movie.Companion.DUMMY_MOVIES.map {
+            MovieListItem.MovieItem(
+                it,
+            )
+        }
 
-    override fun getAdvertisementList(): List<AdItem> = Advertisement.DUMMY_ADS.map { AdItem(it) }
+    override fun getAdvertisementList(): List<MovieListItem.AdItem> =
+        Advertisement.Companion.DUMMY_ADS.map {
+            MovieListItem.AdItem(
+                it,
+            )
+        }
 
     override fun updateMovieList() {
         val movies = getMovieList()
@@ -20,8 +29,8 @@ class MovieListPresenter(
     }
 
     private fun movieListItems(
-        movies: List<MovieItem>,
-        ads: List<AdItem>,
+        movies: List<MovieListItem.MovieItem>,
+        ads: List<MovieListItem.AdItem>,
     ): List<MovieListItem> {
         val adsIterator = ads.iterator()
         val list: List<MovieListItem> =
