+++ app/src/main/res/layout/activity_booking_seat.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="#EAEAEA"
-    tools:context=".view.booking.seat.BookingSeatActivity">
+    tools:context=".ui.seat.BookingSeatActivity">
 
     <LinearLayout
         android:id="@+id/screen"
